Adding basic implementation of job_submitter
This can be used to submit jobs to kubernetes and adds any "extra" stuff to the job this module will use.

Currently that involves adding labels:
 - JobId, JobSetId, Queue, ReadyForCleanup

This allows us to store the job data we'll need to refer to on the pod, without having to hold each job in memory for the entirety of its execution.

Maybe we'll change the approach in the future, but this allows us to be more resilient to executor (module) failures, as kubernetes holds all the state rather than this application