Adding very rough implementation of "background tasks"
These tasks are the main responsibility of the application.

The reason they are separate is:
 - Each can be simpler as just do one thing each
 - They don't really interact with each other, so it is easy to have them run individually
 - They can be run with different frequencies. So we don't have to perform all the operations as frequently as the most frequent task (i.e we can keep the cluster well allocated by checking it each 5 seconds, without having to try and delete all jobs out of the cluster every 5 seconds and instead can be every 30 seconds or whatever)

In the future for optimisation, some of these background tasks may be merged to avoid looking up/calculating similar thing in separate threads.

However for now this keeps it easier to understand and reason about and I don't want to add complexity without knowing it is needed