Implementing empty functions for queue service so executor can call it and get a reply
Currently the queue service panics on all functions. This is inconvenient for testing the executor/api communication is working.

I have changed it to return empty/success for each function so the executor can call it successfully.
The functionality will need to be implemented at some point

Removing commented code from JobLeaseService as it can now call the queue service