Setting entrypoint rather than cmd in dockerfiles
This makes it easier to pass in commandline arguments to docker.

We currently set CMD, which when you run docker, you'll need to say:

docker run ./main <args>

I'd rather people not need to know the internal executable of the docker to make it run.

With ENTRYPOINT this just defines the executable that is run when docker starts

So now you can use it like:

docker run <args>