Adding caching to jobs submitted to Kubernetes
When we submit to kubernetes, in theory there is a latency between submit -> created on server

When we request more jobs, we specify the amount of available resource.
 - Available =  total resource - resource allocated by pods the server

This leaves a gap where a pod can be submitted but not included in the available resource calculation. Meaning we request resource that is in theory already allocated

Now we cache pods between submission and creation on the server. So our available resource calculation is correct