Defaulting last statusChange for pending pods to CreationTimestamp
We know as the pod is in pending, that if no conditions have changed, the last thing that happened must have been creation.

This change is mostly to allow identification of pods that have got stuck in pending but with no events.
 - If a pod gets stuck in pending and never gets an event, currently we'll never delete it

This change just means if it gets truly stuck in pending, we'll clean it up based on being a long time since it was created.