Changing watch to use a context rather than recalculating state summary each time
Currently whenever we want to summary of the events we watch. We loop through every job we have seen events for and sum up which state they are currently in.

This becomes very slow when you have 50000+ jobs, as each event you'll process all 50000 jobs

Now we hold a summary of the current state, so it isn't recalculated on each event.
 - Each event just updates the current state
 - Getting a summary of the current state is just a case of looking it up, rather than recalculating

We could do more optimisations to make starting watching a big jobset faster:
 - Start from current (This may make the summary confusing however, as it won't include all jobs in the jobset)
 - We could not print anything until we catch up with the current state
   - Or just print every 1s rather than every event

Printing takes most of the time now. So skipping past until current saves a LOT of time