When scheduling, check if a job can fit on any nodes available on the cluster (#375)
* When scheduling, we now check if a job can fit on any of the nodes available on the cluster_allocation.go

The executor now reports the largest nodes it has available:
 - Largest meaning the largest of each shape.
 - 1cpu, 2gb node is a different shape than 2cpu, 1gb node.Neither larger than the other

When scheduling, we check if a job will fit on any of these provided nodes

If it doesn't fit, it won't be scheduled. This is to prevent big jobs being allocated to clusters that cannot ever schedule them

* Fixing unit tests

* Removing local testing code

* Replacing IsGreaterThan with Dominates

If a ComputeResource dominates another, it means all it values are >= then the other. With at least once value which is strictly > than the other.

* Renaming getNodesWithAllocatableResource to getAllocatableResourceByNode

It looked like it was filtering the nodes. Instead it should just be providing resource per node.

Also it no longer filters "invalid" nodes and just returns 0 resource allocatable for these nodes.