Checking jobs meet the minimumJobSize when scheduling them
Jobs should be at least as big as minimumJobSize requirements of the cluster to be allowed to be scheduled on that cluster.

Currently we use this feature to make sure only GPU jobs go to GPU clusters. However as we aren't filtering jobs on job size when scheduling, we are seeing CPU jobs be sent to GPU clusters.