Fixed incorrect jobfile example in design overview (#794)
Previously, it was possible to submit a jobfile consisting of just a podspec + queue name and priority, but this option has since been deprecated. Now, jobs must always be given as a list of jobs. This PR changes an example using the old (deprecated) format to use the new format.