Fix unit tests on macOS (#1047)
Docker Desktop on macOS doesn't give connectivity when '--network=host'
is specified, so omit that when running on Macs. Also, the startup for
Redis and Postgres containers can take a couple seconds, and early tests
may fail, so sleep a few seconds before invoking tests. Put `tearDown`
shell func declaration on one line, to satisfy Make and/or Bash on
macOS.

For the test failures, on macOS, the CPU resources returned appears to
be microscopically different (e.g. 1.9999999999999998 instead of 2.0),
(perhaps due to slight math library differences on the platform?) so use
assert.InDelta() with a tolerance of 0.001 to determine if they are
"equal".