Fix 404 errors in api.md (#1172)
The API documentation points to paths relative to github that only
work from github. Instead, we make them fully qualified paths. This
has a side effect of making links from docs/ in forks point to the
upstream official documentation.

Additionally, the helm chart documentation has been consolidated into
a single file to permit it to be integrated into armadaproject.io.
This was required to get rid of all 404 errors in /api since it links
to the helm charts.

In the future, we may want to consider setting up a helm collection
in gh-pages that can be populated via a directory, but given the helm
docs hadn't been touched in almost 2 years, the simpler solution seemed
wiser.

Issue: #889

Co-authored-by: Dejan Zele Pejchev <pejcev.dejan@gmail.com>