Undo commit f263b353 (run Docker as host user uid/gid)
The previous strategy of running the Docker container as the uid/gid of
the container host user (invoking Make) doesn't currently work in our
CICD setup - the volume mount for the Go compiler cache does not match
the uid/gid, so compilations fail immediately. Fixing this will be
considerably more work and testing, so for now, just use the
previous-previous strategy of doing a chmod on the generated
proto/google/protobuf/compiler directory, so subsequent runs can remove
it. Also, add a 'dotnet-clean' target to remove ancillary .NET generated
artifacts.