Remove failed to match logging during scheduling
This line gets logged a lot and doesn't actually mean anything useful is wrong

If there a cluster with 2 node types (A, B).

Jobs that require node type A will cause this log line to fire every scheduling round when node type A is full

This isn't particularly useful to know and definitely isn't an error. So removing this log line