Armada integration/e2e tests in Github Runners (#1156)
Add integration tests that utilize buildjet runners.

Jobs run in context of G-Research will use the paid runner in the Go End To End (buildjet) workflow. These have a runtime of 10-15 minutes. 
Jobs run in context of a fork will use a free runner in the Go End To End (free) workflow. These have a runtime of 15-20 minutes or more.