Expand locking to encompass entire SQLJobService.UpdateJobServiceDb func
This solves the flaky unit test TestConcurrentJobStatusUpdating which
should, in turn, indicate this is a more robust way for
UpdateJobServiceDb to operate. However, I'm not very happy with this
because I don't understand *why* its better. I'm guessing it has
something to do with including the `PrepareDb` call to sqlite, but I'm
not certain.