Add UnmarshalJSON() implementation for EventSequence_Event
This is necessary because EventSequence_Event has its 'Event' member
which is an interface (isEventSequence_Event_Event), which is not
directly supported for unmarshaling. The solution is to use a temporary
struct that has Event as a json.RawMessage, and then we attempt to
unmarshal that Event (as a RawMessage) to each of the possible
underlying structs for the isEventSequence_Event_Event interface.
This can obviously get slow, but it seems to be reasonably robust.