Improve logic for an EventSequence_Event struct
That struct has a member 'Event' which is an interface - unmarshal it
into a json.RawMessage, and also unmarshal it into an emphemeral
map[string]interface{}, so we can reliably get the JSON key for the
underlying struct and switch on that for the correct json.Unmarshal()
against the real struct underneath. This logic now uses just two
invocations of json.Unmarshal().

Add the start of a unit test for the UnmarshalJSON method - more
structs that implement the EventSequence_Event.Event interface
will be added.

Change (Pulsar) Producer name format per PR feedback.