WIP of a command-line client to generate Pulsar msgs (#1507) (#1559)
* WIP of a command-line client to generate Pulsar msgs (#1507)

To build: use the `build-pulsartest` makefile target.

Example usage:

  bin/pulsartest --jobsetEventsTopic <topic> submit <events yaml>
e.g.
  bin/pulsartest --jobsetEventsTopic armada/armada/events-partition-0 submit eventsequence.yaml

where eventsequence.yaml is a serialized armadaevents.EventSequence object

* Send event-sequence synchronously to Pulsar.

Remove deferred Pulsar Producer/Client closing.

* - Remove nascent dry-run option - it's not needed....

- Remove PulsarSubmitServer member - we are just using a Pulsar
  Producer directly.

- Remove unused app.Params - it wasn't really used, since we just
  immediately construct a Pulsar Producer object from the given
  command-line flag values.

- Remove Producer.Flush() invocation - it's not needed for a synchronous Send().

- Remove unused LoadCommandlineArgsFromConfigFile() - we will just use
  command-line options.

- Re-run goimports with '-local' option for G-Research packages.

* Add missing build package; process YAML as single doc.

Add missing build package for versioning info in resultant binary.
Also, simplify YAML process - the incoming YAML file is not
multiple conjoined docs, just process it as a single doc.

* Add UnmarshalJSON() implementation for EventSequence_Event

This is necessary because EventSequence_Event has its 'Event' member
which is an interface (isEventSequence_Event_Event), which is not
directly supported for unmarshaling. The solution is to use a temporary
struct that has Event as a json.RawMessage, and then we attempt to
unmarshal that Event (as a RawMessage) to each of the possible
underlying structs for the isEventSequence_Event_Event interface.
This can obviously get slow, but it seems to be reasonably robust.

* Remove debugging message.

* Avoid null-ptr dereferences when incoming request is incomplete.

If the incoming request is missing a child KubernetesMainObject
(or if the whole EventSequence itself is a null pointer), return
an immediate error. This was found while testing with a SubmitJob
that was not completely filled in.

* Improve logic for an EventSequence_Event struct

That struct has a member 'Event' which is an interface - unmarshal it
into a json.RawMessage, and also unmarshal it into an emphemeral
map[string]interface{}, so we can reliably get the JSON key for the
underlying struct and switch on that for the correct json.Unmarshal()
against the real struct underneath. This logic now uses just two
invocations of json.Unmarshal().

Add the start of a unit test for the UnmarshalJSON method - more
structs that implement the EventSequence_Event.Event interface
will be added.

Change (Pulsar) Producer name format per PR feedback.

* golangci-lint demands 'marshalled'.

* Add UnmarshalJSON() implementations for some K8S types

Implement custom UnmarshalJSON() methods for KubernetesMainObject,
KubernetesObject, and KubernetesResourceInfo structs, as each of these
has an interface member (which are not natively unmarshal-able).
First, unmarshal into a map, then use the key to determine the
underlying concrete struct, and do a real unmarshal thereafter.

* Add more unmarshaling test logic.

Add more comprehensive testing of an EventSequence_Event as well as a
full EventSequence object. Move private func expectedSequence-
FromRequestItem() from pulsar_test.go to a public ExpectedSequence-
FromRequestItem() into pkg/armadaevents/ events_util.go, so we can use
it to also generate a full-featured EventSequence to more fully test the
unmarshaling.

* Remove unused, vestigial App.Out stream member.