Don't report status changes on pods being deleted by Armada
Currently when you cancel a job that has a non zero terminationGracePeriodSeconds, we get these events
```
JobSubmittedEvent
JobQueuedEvent
JobLeasedEvent
JobPendingEvent
JobRunningEvent
JobCancellingEvent
JobCancelledEvent
JobTerminatedEvent
JobFailedEvent
```

We shouldn't get the JobFailedEvent as this is a second terminal event.

We didn't get these in the past as we used to set terminationGracePeriodSeconds to 0 for all pods, which meant kubernetes didn't send the state transitions after we called delete.

To avoid this, we should just simply skip sending status updates on pods Armada is trying to delete