Split out GH Actions workflow for pushing release images
Remove the logic that uploads Docker images to GH Artifacts, downloads
the images and pushes to Docker Hub from go-integration.yml, as that
workflow runs on all branches *except* for master. We want to do the
image-push to Docker Hub only on master, so introduce the new
build-release-images.yml to do this, with the 'on' event restricted to a
push to master.