Save Docker images to GH Artifacts and push to Docker Hub (#1272) (#1682)
* Save built Docker images to Github Artifacts.

They will be used in a subsequent job of the workflow to
upload to Docker Hub.

* Push Docker images from GH Artifacts to Docker Hub.

Also, update GHA upload-artifacts plugin to latest version.

* Do not look for a zipfile - the bundle is a gzipped tarfile.

* Revert "Do not look for a zipfile - the bundle is a gzipped tarfile."

This reverts commit 3529124fd61f73bbd5a0f7aa9d47cb56d2348ff0.

* Add dir prefix to zipfile when unzipping.

* Try different path component spec.

* Add temporary find invocation, to locate zip file.

* Attempt different find for locating images.

* Skip unzipping, and just untar the download.

* Remove commented-out Actions debugging directives

* Push to new 'gresearch' org in Docker Hub.

* Split out GH Actions workflow for pushing release images

Remove the logic that uploads Docker images to GH Artifacts, downloads
the images and pushes to Docker Hub from go-integration.yml, as that
workflow runs on all branches *except* for master. We want to do the
image-push to Docker Hub only on master, so introduce the new
build-release-images.yml to do this, with the 'on' event restricted to a
push to master.