In 'tests' Makefile target, run separate tests for Job Svc
Run separate tests for the Job Service repository, once using Postgres,
and once using SQLite, so we test both db types in each build run.
Rename each test func in the internal/jobservice/ repository package to
have prefix of 'TestJobSvc', so that we can use the '-skip' option to
'go test' to ignore them, and then we run the tests on those separately.