Use connection pooling for Postgres in Job Service database (BATCH-170) (#2378)
* Start a Postgres container for tests-e2e-airflow Make target.

* Revert "Start a Postgres container for tests-e2e-airflow Make target."

This reverts commit b9b406c960de2fd12a5075d683591a93725ad827.

* Use pgx connection pooling on Postgres (BATCH-170)

Use the pgx.Pool facility to use connection pooling for the
Job Service database when configured to use Postgres for the
JS service. Split out the Postgresql and SQLite implementations
and mask behind a common SQLJobService interface definition.
Currently, all unit tests function correctly when using Postgres.

* Get Postgres and SQLite configs working in unit tests

* Add new context param to various func invocations.

* Also remove the test.db-shm and test.db-wal test files for SQLite

* golangci-lint fixes

* Add short sleep for CI Integration tests

To avoid the "no executors available" error, when containers
are not fully up.

* Try a 2-minute sleep before running e2e tests after cluster build.

* Remove debugging comment.

* Add error object to NewSQLJobService() constructor

Improve error-handling when constructing new database connection,
do not just log errors, but pass them up to callers. Render a
useful error message if the database type is not valid. Thanks
to @suprjinx for PR suggestions.