Job Service: Rework logic for jobset subscriptions (BATCH-303)
Rework the logic for how we do processing of jobset subscriptions:
change from launching a new goroutine to process each subscription
(in the set we retrieve from the database), to feeding them into a
channel, which will be drained by a fixed set of goroutines in
a pool.

Also, change the UNIQUE constraint in the jobsets table to a PRIMARY
key (to forbid nulls).