Show job run errors below container logs in details (#3032)
In the job details pop-up, in the LOGS tab, if there are job run errors,
render them below the div showing the container logs.

This is intended to be most useful when the container was not even able
to start - quite often when an incorrect image name/version was
specified, and the system was unable to pull the image. Without this,
the container logs div just shows an error that it was unable to pull
any logs from Binoculars.

Signed-off-by: Rich Scott <richscott@sent.com>