Return more detailed submission/validation errors.
Generate and return more detailed submission and/or validation errors.
If there are numerous jobs with errors, just give the number of failed
jobs (and the total number originally submitted), and truncate the list
of failed jobs errors to just the first 5 (this is defined in a single
constant variable, if neededed to change later).

Signed-off-by: Rich Scott <richscott@sent.com>