If ARMADA_COMPONENTS env var is empty, return zero-sized slice
This fixes a bug where, if the ARMADA_COMPONENTS env var was a
zero-length string (or simply unset),  getComponents() would return a
slice of length 1, where the entry was a zero-length string. This was
causing errors when subsequent conditional to see if the generated slice
was zero-length wasn't getting triggered.

Signed-off-by: Rich Scott <richscott@sent.com>