Use separate db for jobservice integration tests (#521) (#3373)
Create and use a separate database for the Job-Service integration tests
(known as the `teste2eAirflow` Mage target). The localdev Armada setup
uses the default/built-in database 'postgres', and since there are
table-name collisions between the Armada and the Job Service databases
(e.g. both have a 'jobs' table), create and use a db named 'jobservice'
for the Job Service integration tests.

Signed-off-by: Rich Scott <richscott@sent.com>