Fixed broken tests by fixing unfeasibleSchedulingKeys bug
We had a bug where unfeasibleSchedulingKeys was being populated incorrectly, which led to all jobs of a particular shape being banned for that scheduling round. By fixing this bug we also broke some tests which relied on UnsuccessfulJobSchedulingContexts being populated correctly. However, this data structure was only populated if a job was marked as unsuccessful because it's scheduling key was in unfeasibleSchedulingKeys. Fixing this bug should also fix inaccuracies in metrics, as well as provide better job scheduling reporting.