Add copy buttons to lookout UI
Add copy icon buttons, which appear on hover, next to appropriate values in the jobs table and in the key-value pair rows in the details sidebar. Also add a copy button for the job ID at the top of the sidebar, which always remains visible.

When the user clicks the button, a tooltip displays *Copied!* to confirm to the user that the value has been copied to their clipboard.