Use virtualized table for job sets table (#4116)
We [previously](https://github.com/armadaproject/armada/commit/a57e7f2622a4efb8e70b24415ec75046ecd5824d) made the job sets table a normal, non-virtualized table. This was to simplify the logic, remove the third-party library, `react-virtualized`, which is no longer well-maintained, and done on the assumption that the page would not have to display enough rows of data to cause it any performance-related trouble.

We have observed in some environments that the latter assumption is not correct, and have noticed significant performance degradation with very large datasets.

This commit makes the table a virtualised one again, but using the well-maintained Virtuoso library. It also stops truncating the job set ID as users have indicated it is often useful to see the whole thing, even if it is long.

Signed-off-by: Rich Scott <richscott@sent.com>