Allow configuration of column order (#4118)
In the lookout UI, allow the user to change the order of the columns of the jobs table through a drag-and-drop interface.

This change replaces the existing column selector with a new column configuration dialog which extends the current functionality of the column selector with the ability to drag and drop columns into the user's desired order.

The existing functionality which has been preserved is:
	- Hiding and showing columns (via a checkbox)
	- Adding annotation key columns
	- Removing and editing annotation key columns which have been added
	- Indicates when columns are for annotation keys

The new column configuration dialog has the following additional functionalities:
	- Hides columns which are configured to be pinned (i.e. the selector column)
	- Hides columns which are grouped, and explains this to the user
	- When the input to add an annotation column is not visible on the screen, it displays a chip indicating its presence. Clicking on this chip will scroll to the input
	- Validates annotation key columns on input, making sure they have no leading or trailing whitespace, and that there is not already a column for the same annotation key

This change also makes the buttons in the job action bar more visually consistent - using primary colours for actions and secondary colours for configuration changes.

Signed-off-by: Rich Scott <richscott@sent.com>