Visual improvements to the job sidebar (#4133)
This change makes the sidebar for viewing a job's information visually more consistent across its tabs and the rest of the application. We also migrate to using styled components using emotion JS, which is provided by MUI, instead of separate CSS files. This provides better type-safety and keeps styles in the same place as the components. Note that this has necessitated an upgrade to vitest, and the installation and use of @emotion/babel-plugin.