Add scheduling reports to Lookout UI
In the lookout UI, add a new sidebar tab, "Scheduling", when viewing a job in the queued state. This tab displays the time the job has spent in the queue, scheduling reports for the job (reports for the queue and overall are also available, but initially hidden), and parameters which are significant in the context of scheduling - the priority and resource requests.

This tab aims to help users work out why their job is not scheduling.