Add scheduling reports to Lookout UI (#4144)
Add scheduling reports to a new sidebar tab for jobs which in a queuing state. This tab displays the time the job has spent in the queue, scheduling reports for the job (reports for the queue and overall are also available, but initially hidden), and parameters which are significant in the context of scheduling - the priority and resource requests.