Prevent multiple OIDC authn redirects on Lookout UI (#4196)
This fixes race conditions in the Lookout UI which cause repeated and unnecessary redirects when authenticating the user using OIDC.

We also add loading and error screens which are displayed during the sign-in process. On the happy path, the loading screen is displayed very briefly only when (re-)authentication is required.