Autocomplete filter for queue names (#4184)
Add autocompletion for the filter for queues on the job sets page and the queue column filter in the jobs table.

This change means that database queries which include queue filtering on are much more efficient - `WHERE queue IN ('myqueue', 'otherqueue)` rather than `WHERE queue LIKE 'myqueue%'`. This is because the query planner can use one of the existing (multicolumn) indexes on queue.

On the front end, we use the MUI Autocomplete component and the existing endpoint to get queues. The response of this endpoint is a JSON stream, which is not currently supported by our OpenAPI/TS setup, hence the JSON string transformations in `internal/lookout/ui/src/services/lookoutV2/useGetQueues.ts`.

We also add the `msw` to easily mock fetch responses in unit tests. We wish to use this across all tests in future, rather than mocking our custom services (which means the unit tests do not ensure that our actual fetch requests are correctly formed).

We add a mechanism to migrate filter values where the match-type of the filter has changed to or from any-of, as has happened here to the queue filter:
- For a column whose match type is any-of, if its filter value is not an array, we clear the filter completely. This prevents user confusing if their stored value is not for an exact match - they will expect to see results which are no longer displayed.
- For a column whose match type is **not** any-of, if its filter value is an array, we set the filter value to the first element of the array.