Show the user who cancelled a job in Lookout (#4236)
Add a field to the CancelledJob message which records the user who initiated the request to cancel a job. In the case of multiple users cancelling the same job, we record the first user who made the request.

In the scheduler, we add a column to its database for the cancel user and in its ingester, we store the requesting user upon receiving a CancelJob or CancelJobSet. The scheduler publishes a JobCancelled message with new field correctly populated.

In the Lookout ingester, we read this field's value and store it in a new column in the Lookout database when a job has been cancelled.

We then display this on the Lookout UI if it is populated.