Lookout UI: fix selection of jobs involving groups (#4238)
This fixes a bug in the Lookout UI. When a user uses grouping in the jobs table, if they select a group but then de-select rows within that group, the de-selection is no respected when using it to perform actions. This is because the job filters do not exclude the de-selected rows when executing the query. We fix this by recursing down the tree of grouped rows (see getFiltersForRowsSelection() in jobTableUtils.ts).

We also re-enable fetching more rows in a group when there are more than the selected page size in the group.