Add typecheck command for Lookout UI (#4240)
This change introduces a new yarn typecheck command which simply runs tsc. The README for the Lookout UI has been updated to encourage developers to use this command during development. The command is also run in the existing Tests GitHub workflow, which will therefore now fail if there are typechecker errors.