Ensure cache-control headers are always served for index.html response
In the Lookout UI web server, the HTTP response headers preventing index.html from being served are only served when the document is accessed directly i.e. when navigating to "/" or "/index.html". They are not served when the document is returned in response to a request path which doesn't exist. (This behaviour is needed for single-page applications because the handling of 404s is deferred to the front end's client side routing.)

This commit makes sure these headers are served for the latter as well as the former.