Fix: prevent overwriting of pagination and sidebar job parameters in Lookout UI (#4259)
This fixes a bug in the Jobs table on the Lookout UI where URL
parameters are overwritten when the table is initially rendered. The
affected parameters are for pagination and the job sidebar.

The cause of the bug is the `onFilterChange()` function is called on the
initial render of the table, and this function performs side-effects
which wipe these parameters. These side effects should only be called
when the column filters have materially changed. This fix adds a check
for this.