Apply 'exactly' match-type when clicking filter button (#4260)
In the Lookout UI's jobs table, users can filter by the value in a cell
by clicking on a button next to the value which appears on hover, so
long as the column supports filtering.

Currently, for non-enum values, the filter value is applied, but the
match-type is left unchanged (this defaults to starts-with). This
changes the match-type to 'exactly', which is both more appropriate and
more efficient for the database to process.