Fix text filter pasting behaviour (#4300)
When pasting into a text filter in the Lookout UI, we trim the pasted text to remove whitespace at the start or end. Currently, the pasted text replaces the entire contents of the text box. This fix prevents the existing contents of the text box from being overwritten - rahter, it respects the cursor any selected text.