Add download button for logs in Lookout (#4297)
Similar to how users can download the YAML spec of a job, this commit adds a button to the logs viewer for users to download the logs displayed, for a particular container for a particular job run.

The text file's name takes the pattern `[job ID]-[run index]-[container name].txt`, for example, '01jpn0zj58rr1898ym3z93t01e-0-alpha.txt`.

Each log line will be preceded by the ISO timestamp if and only if theuser has enabled this with the toggle switch.