Preserve URL path and search params in login flow (#4327)
In Lookout, when the user is redirected to the OIDC provider to log in, store the URL path and search params from which they were redirected, then navigate to it when they return to the app.

This means that when the user follows a link to the jobs page with filters encoded in their URL, if they need to login, they will land back on their desired page after logging in, rather than the default page (which will use their previous set filters from their browser's local storage, if present).