Add error boundaries to Lookout UI
There should never be any uncaught JavaScript errors in the Lookout UI React app. However, in a non-trivial app with ongoing development, it is probably inevitable that these will appear from time to time.

This change adds error boundaries to the application, so that such errors are caught and displayed to the user gracefully, rather than crashing the application (and the user seeing a blank white screen).