Add error boundaries to Lookout UI (#4332)
There should never be any uncaught JavaScript errors in the Lookout UI
React app. However, in a non-trivial app with ongoing development, it is
probably inevitable that these will appear from time to time.

This change adds error boundaries to the application, so that such
errors are caught and displayed to the user gracefully, rather than
crashing the application (and the user seeing a blank white screen).

## Example screenshots (the errors have been manually thrown!)


![image](https://github.com/user-attachments/assets/ac88a56c-f825-4ed2-9500-e405357efcda)

![image](https://github.com/user-attachments/assets/ea1f3126-045e-4cdf-bd14-a8ed13caefab)

![image](https://github.com/user-attachments/assets/6dcdd5b0-150f-4730-9b62-3726b606d9d1)

![image](https://github.com/user-attachments/assets/1e18d7d6-677d-4879-94e6-4bbe3f8c8feb)

![image](https://github.com/user-attachments/assets/43dc3280-dfaf-45a2-8fa1-fb9a02da77a1)