Lookout UI - migrate old links with queue filters
The change to the queue filter in the Lookout UI - from a single-value
filter to a multi-value one-of filter - has made previously generated
URL links unusable. This change migrates URLs and local storage values
which include a queue filter, converting them to the new type.

If this filter's match type is not "exact", i.e. it is a "starts-with"
or "contains", we remove the queue filter completely. The filter's
match-type is now a "one-of" and so jobs expected to match with
non-exact match types will not appear.