Add /jobs/{job ID} redirect links to Lookout (#4352)
Links following this pattern will link to the jobs page, with the single
filter - job id equals {job ID} - applied, and the job details sidebar
open.

This also adds a "share" icon button to the top of the job sidebar,
which copies this link to the user's clipboard for the open job.

This commit additionally centralises the definition of pathnames across
the app into a single file as the single source of truth for these.