Lookout UI (fix location of job run error and debug message)
For a job which has one or more errored runs, we display each errored run's logs and its debug message from Armada if present, in the *Result* tab of the sidebar. The order of these errors and debug messages is not consistent with the order of the runs. This fixes this discrepancy.

This commit also performs some refactoring of the component where this is displayed, as well as the mechanism to execute the fetch requests to obtain the data.