Improve spot price calculation when there are cordoned/stale clusters  (#4389)
Currently we calculate the spot price when we've scheduled 90% of the
available resource

However when clusters are cordoned/not heart beating, the total resource
is lowered to remove these clusters however the allocated resource still
reflects the resource on these clusters.

Before:

cluster 1 - 100 cpu (cordoned)
cluster 2 - 100 cpu

Assuming both clusters are fully allocated, we'll start scheduling based
on:

total resource - 100
spot price threshold - 90
allocated - 100

So the first job on cluster 2 will determine the price, as we're already
past the threshold.

This PR:
We will only consider resource scheduled in this round towards the
threshold, resource on cordoned/lagging clusters will be ignored for
determining the spot price

---------

Signed-off-by: JamesMurkin <jamesmurkin@hotmail.com>